requirements.txt
# 中国福利彩票预测工具依赖包

# 数据处理
pandas>=1.3.0
numpy>=1.21.0

# 机器学习
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0

# 深度学习（可选）
# tensorflow>=2.8.0
# torch>=1.10.0

# 配置和工具
PyYAML>=6.0
joblib>=1.1.0
click>=8.0.0

# 数据可视化（可选）
matplotlib>=3.5.0
seaborn>=0.11.0

# 其他工具
tqdm>=4.62.0  # 进度条
requests>=2.26.0  # 网络请求（用于数据爬取）
beautifulsoup4>=4.10.0  # 网页解析（用于数据爬取）